Added tabs to profile
