+++ views/show/view.php
@@ -1,8 +1,6 @@
 <?php
-
-use app\helpers\SkillsHelper;
+use kartik\tabs\TabsX;
 use app\helpers\ViewHelper;
-use yii\helpers\Html;
 
 /* @var $this yii\web\View */
 /* @var $model app\models\Show */
@@ -51,68 +49,50 @@ $this->params['breadcrumbs'][] = strtok($this->title, " ");
                 </div>
             </div>
         </div>
+        <?php
+
+        $projects = $this->render('_projects', [
+                'model' => $model,
+                'switch' => $switch,
+                'projects' => $projects,
+                'urlHelperForProjects' => $urlHelperForProjects,
+                'course' => $course,
+                'parents' => $parents,
+                'projects' => $projects,
+        ]);
+
+        $tmp = Yii::$app->session->get('username');
+
+        $items = [
+            [
+                'label'   => '<i class="glyphicon glyphicon-list"></i> ' . Yii::t('app', 'Projects'),
+                'content' => $projects,
+                'active'  => true
+            ],
+            [
+                'label'   => '<i class="glyphicon glyphicon-time"></i> ' . Yii::t('app', 'Time in cluster'),
+                'content' => '',
+            ],
+            [
+                'label'   => '<i class="glyphicon glyphicon-ok-circle"></i> ' . Yii::t('app', 'Corrections Log'),
+                'content' => '',
+            ],
+            [
+                'label'   => '<i class="glyphicon glyphicon-gift"></i> ' . Yii::t('app', 'Achievements'),
+                'content' => '',
+            ],
+        ];
+?>
         <div class="col-lg-9 mx-auto">
             <div class="card" style="width: 100%;">
                 <div class="card-body">
-                    <h5 class="card-title">Projects</h5>
-                    <div style="max-width: 100%; overflow: auto;">
-                        <table class="table ">
-                            <thead>
-                            <tr>
-                                <th scope="col">name</th>
-                                <th scope="col">final mark</th>
-                                <th scope="col">Retry</th>
-                                <th scope="col">status</th>
-                            </tr>
-                            </thead>
-                            <tbody id="alldata">
-
-                            <?php if (isset($projects)) { foreach ($projects as $item) {
-                                if ($item->name === 'Rushes') { continue; } ?>
-                            <tr>
-                                <td>
-                                    <div class="progress my-shadow">
-                                        <div class="progress-bar progress-bar-<?= ViewHelper::getProgressProjectColor($item->final_mark, $course, $item->status)?>" role="progressbar" style="width:<?= ViewHelper::getProgressProject($item->final_mark, $course)?>%">
-                                            <p><a style="color:black; " href="<?= '/' . Yii::$app->language . $urlHelperForProjects ?><?= $item->slug ?>"><?= $item->name ?></a></p>
-                                        </div>
-                                    </div>
-                                </td>
-                                <td><?= $item->final_mark ?></td>
-                                <td><?= $item->occurrence ?></td>
-                                <td><?= str_replace('_', ' ', $item->status) ?></td>
-                            </tr>
-                            <?php } } ?>
-
-
-                            <?php if (isset($parents)) { foreach ($parents as $parent => $value) { ?>
-                                <tr><td colspan="4"><h5 class="card-title"><?= $parent?></h5></td></tr>
-                            <?php if (isset($value)) { foreach ($value as $item) { ?>
-                                <tr>
-                                    <td>
-                                        <div class="progress my-shadow">
-                                            <div class="progress-bar progress-bar-<?= ViewHelper::getProgressProjectColor($item->final_mark, $course, $item->status)?>" role="progressbar" style="width:<?= ViewHelper::getProgressProject($item->final_mark, $course)?>%">
-                                                <p><a style="color:black; " href="<?= '/' . Yii::$app->language . $urlHelperForProjects ?><?= $item->slug ?>"><?= $item->name ?></a></p>
-                                            </div>
-                                        </div>
-                                    </td>
-                                    <td><?= $item->final_mark ?></td>
-                                    <td><?= $item->occurrence ?></td>
-                                    <td><?= str_replace('_', ' ', $item->status) ?></td>
-                                </tr>
-                            <?php } } } } ?>
-
-                            </tbody>
-                        </table>
-                    </div>
+                <?php
+                echo TabsX::widget([
+                    'items'=>$items,
+                    'position'=>TabsX::POS_ABOVE,
+                    'encodeLabels'=>false
+                ]);
+                ?>
                 </div>
-                <hr>
-                <!-- @TODO need create user time in clusters method -->
-                <!--            <div class="card" style="width: 100%;">-->
-                <!--                <div class="card-body">-->
-<!--                    <h5 class="card-title">Time at cluster</h5>-->
-<!--                    <canvas id="timer" width="770" height="385" style="display: block; width: 770px; height: 385px;"></canvas>-->
-<!--                </div>-->
-<!--            </div>-->
-            </div>
             </div>
         </div>
